Defend against eronious newlines in smartdown markdown values
They may contain additional leading/trailing whitespace and/or new lines, which
conflicts during the diffing against an existing Smart Answer

It may also return a single new line, which was treated as true when inverted, so
use rails present? instead.

This needs fixing in Smartdown, or at least documented in a spec to show the
expectations around markup whitespace so a client, like this app, knows what
it needs to handle